refactor: shared component and utilities in advance of route details
