+++ androidApp/src/androidTest/java/com/mbta/tid/mbta_app/android/component/DirectionPickerTest.kt
@@ -8,7 +8,6 @@ import androidx.compose.ui.test.performClick
 import com.mbta.tid.mbta_app.android.stopDetails.DirectionPicker
 import com.mbta.tid.mbta_app.model.Direction
 import com.mbta.tid.mbta_app.model.ObjectCollectionBuilder
-import com.mbta.tid.mbta_app.model.StopDetailsFilter
 import junit.framework.TestCase.assertTrue
 import org.junit.Rule
 import org.junit.Test
@@ -24,22 +23,20 @@ class DirectionPickerTest {
                 color = "000000"
                 textColor = "ffffff"
             }
-        var filter: StopDetailsFilter? = StopDetailsFilter(routeId = route.id, directionId = 0)
+        var directionId = 0
         composeTestRule.setContent {
             DirectionPicker(
                 availableDirections = listOf(0, 1),
                 directions = listOf(Direction("North", null, 0), Direction("South", null, 1)),
                 route = route,
-                line = null,
-                filter = filter,
-                updateStopFilter = { newFilter -> filter = newFilter },
+                selectedDirectionId = directionId,
+                updateDirectionId = { directionId = it },
             )
         }
 
         composeTestRule.onNodeWithText("Northbound").assertIsDisplayed()
         composeTestRule.onNodeWithText("Southbound").performClick()
-        assertTrue(filter?.routeId == route.id)
-        assertTrue(filter?.directionId == 1)
+        assertTrue(directionId == 1)
     }
 
     @Test
@@ -50,16 +47,15 @@ class DirectionPickerTest {
                 color = "000000"
                 textColor = "ffffff"
             }
-        var filter: StopDetailsFilter? = StopDetailsFilter(routeId = route.id, directionId = 1)
+        var directionId = 1
         composeTestRule.setContent {
             DirectionPicker(
                 availableDirections = listOf(1),
                 directions =
                     listOf(Direction("North", null, 0), Direction("South", "Destination", 1)),
                 route = route,
-                line = null,
-                filter = filter,
-                updateStopFilter = { newFilter -> filter = newFilter },
+                selectedDirectionId = directionId,
+                updateDirectionId = { directionId = it },
             )
         }
