+++ shared/src/commonMain/kotlin/com/mbta/tid/mbta_app/model/Direction.kt
@@ -3,7 +3,7 @@ package com.mbta.tid.mbta_app.model
 import co.touchlab.skie.configuration.annotations.DefaultArgumentInterop
 import com.mbta.tid.mbta_app.model.response.GlobalResponse
 
-data class Direction(var name: String, var destination: String?, var id: Int) {
+data class Direction(var name: String?, var destination: String?, var id: Int) {
     /**
      * This constructor is used to provide additional context to a Direction to allow for overriding
      * the destination label in cases where a route or line has branching. We want to display a
