+++ shared/src/commonMain/kotlin/com/mbta/tid/mbta_app/model/RouteCardData.kt
@@ -539,6 +539,13 @@ data class RouteCardData(
                     is Line -> this.routes.min()
                 }
 
+        val allRoutes: Set<RouteModel>
+            get() =
+                when (this) {
+                    is Route -> setOf(this.route)
+                    is Line -> this.routes
+                }
+
         fun directions(
             globalData: GlobalResponse,
             stop: Stop,
